[ISSUE-1254] Move test contracts to API
