+++ tmail-backend/jmap/extensions-api/src/test/scala/com/linagora/tmail/james/jmap/label/LabelChangeRepositoryContract.scala
@@ -344,20 +344,3 @@ trait LabelChangeRepositoryContract {
   }
 
 }
-
-class MemoryLabelChangeRepositoryTest extends LabelChangeRepositoryContract {
-  var repository: MemoryLabelChangeRepository = _
-  var updatableTickingClock: UpdatableTickingClock = _
-
-  override def testee: LabelChangeRepository = repository
-
-  override def stateFactory: State.Factory = State.Factory.DEFAULT
-
-  override def setClock(newTime: ZonedDateTime): Unit = updatableTickingClock.setInstant(newTime.toInstant)
-
-  @BeforeEach
-  def setup(): Unit = {
-    updatableTickingClock = new UpdatableTickingClock(DATE.toInstant)
-    repository = MemoryLabelChangeRepository(updatableTickingClock)
-  }
-}
