+++ tmail-backend/jmap/extensions-api/src/test/scala/com/linagora/tmail/james/jmap/settings/JmapSettingsRepositoryContract.scala
