+++ tmail-backend/jmap/extensions-cassandra/pom.xml
@@ -23,16 +23,11 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>jmap-extensions</artifactId>
+            <artifactId>jmap-extensions-api</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>jmap-extensions</artifactId>
-            <scope>test</scope>
-        </dependency>
-
         <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>apache-james-backends-cassandra</artifactId>
